Initial commit
This is the initial commit of the Rust bughunt project. The project's
aim is to provide a suite of QuickCheck tests for Rust's standard
library data structures with the ambition, as stated, of hunting for
bugs. This work is inspired by the recent fuzzing work of Sergey
"Shnatsel" Davidoff in Rust and prior work applying QuickCheck methods
to Erlang by Jesper L. Andersen. Much thanks to Anthony Ramine for
giving me the final bit of inspiration needed to get this project
rolling.

This commit targets `std::collections::HashMap`, somewhat in homage to
the work done by Andersen and also because I could recall HashMap's
interface from memory while flying back from RustConf 2018.

The actual test code is fairly minimal at this point. The skeleton has
been placed, in this commit. Now it needs flesh on its bones.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>