Finish up 'intro' HashMap
This commit puts the `PropHashMap` implementation to where I wanted
it to be for introducing the project. I've implemented a hasher that
is very likely to cause collisions -- the TrulyAwfulHasher -- and
have got docs spread around. The actual coverage on HashMap's
interface is pretty poor and we still have some deeper semantics
around keys to tackle, but the idea should be plain enough.

HashMap is fairly well used. I'll be surprised if I turn up any bugs
in it.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>