Introduce an initial travis CI configuration (#3)
* Introduce an initial travis CI configuration

This commit introduces a basic Travis CI configuration into the
project. QC tests will just fly right on by but our clippy and
formatting warnings will be checked. The developer should still
run their models on their local machines for long periods.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Introduce Travis CI shield

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Adjust call to cargo fmt

I _believe_ that I was passing older-style arguments to rustfmt.
This commit is now based on the flags taken by current rustfmt.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Satisfy clippy

This commit has two changes, placing QuickCheck in ticks to offset
it in the documentation and a complexity change, removing 'as'
in favor of a call to `u64::from`.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Instruct travis to cache cargo directory

Build times with Rust in CI are not entirely fast. This config option
will instruct the CI system to keep what pre-compiled information it
can between runs.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Disable dual runs in travis CI

I'd misunderstood how implicit builds worked in the travis configuration,
resulting in two nearly identical runs. This doubled the build time.
Oops.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>