Introduce more Ops for HashMap
This commit introduces three new operations against HashMap:
ShrinkToFit, CheckCapacity and Clear. The consequences of these
operations are documented in the commit itself. I've also updated the
travis CI runner to run more tests of larger size by default. This
will need to be dialed backward as the number of models grow.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>