Experiment with debug/release builds
Rust behaves a little differently whether you've built it in debug
mode or release mode. Integer operations come to mind. A wrap-around
in debug mode will crash where, uh, it'll wrap in release. To call out
that QC needs to be run in both release and debug modes this commit
experiments with getting Travis to build both.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>