Introduce more Ops for HashMap (#4)
* Introduce more Ops for HashMap

This commit introduces three new operations against HashMap:
ShrinkToFit, CheckCapacity and Clear. The consequences of these
operations are documented in the commit itself. I've also updated the
travis CI runner to run more tests of larger size by default. This
will need to be dialed backward as the number of models grow.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Experiment with debug/release builds

Rust behaves a little differently whether you've built it in debug
mode or release mode. Integer operations come to mind. A wrap-around
in debug mode will crash where, uh, it'll wrap in release. To call out
that QC needs to be run in both release and debug modes this commit
experiments with getting Travis to build both.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>