Initial work toward using a fuzzer as a QC tool
There's some problems with the available QC tools, at least for the
purposes of this project. Namely:

 * the Rust testing framework limits the amount of concurrency
   available, starving some tests of runtime
 * there is no distribution available
 * test-case generation can be slow

Considering we're randomly searching through a state space the name of
the game is speed. To that end, I have in mind a notion of
QuickChecking which operates first in fuzzer mode--quickly throwing
random nonsense into your program to find crashes, track branch
conditions in the program--and, once a crash has been found, switches
into QC mode to shrink the found case. This commit does not have that
mode operation notion, only shuffles code around to run a property
test under AFL. There is no shrinking.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>