Introduce FiniteBuffer
This commit introduces FiniteBuffer into the hash map
fuzz. FiniteBuffer, unlike RingBuffer, does not produce an infinite
amount of data but will limit production to the total number of bytes
available in `data`. This change means we have to guard each Arbitrary
and it's possible that if the fuzzer is not correctly configured we'll
only check small total operations. But, it is quite a bit faster and
fuzzers that do minimization -- like AFL -- are able to eat into the
shrink step needed for proper quickchecking.

This is promising, I think.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>