Cover `VecDeque::remove`
This commit introduces coverage for `VecDeque::remove`, removing an
element at a specific offset from the VecDeque. If the index is beyond
the bound of the length then None is returned. We _are_ digging too
much into the out-of-bounds case here but I'm not coming up as of this
writing with a clever way to probe both. Maybe introduce a bounding
boolean into Remove, which switches modulo bounding on like we do with
Insert?

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>