Repair and issue with VecDeque Op::Insert, introduce swap_remove_back
This commit repairs and issue with the interpretation of Op::Insert,
namely that it blew up pretty immediately. This snuck through because
I did not properly run the fuzz job, having cranked this thing out
quickly after dinner. I have, now working on a plane, introduce
Op::SwapRemoveBack which tests swap_remove_back. The careful commit
reader will note that the interpretation is commented out. This is a
result of a disagreement between the SUT and the model.

I imagine that the model is at fault but my current tooling makes
debugging this difficult. Need to invest in a shrinking step now.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>