Introduce str_repeat
str_repeat is a fuzz target to test `str::repeat`, the first target in
which we know of an actual bug, described in
https://blog.rust-lang.org/2018/09/21/Security-advisory-for-std.html. We
can now run this against a known-faulty version of Rust and evaluate
fuzzing strategies, similarly for
https://cve.mitre.org/cgi-bin/cvename.cgi?name=%20CVE-2018-1000657. Having
a set of known bugs is a Good Thing per "Evaluating Fuzz Testing" by
Klees et al.

This target illustrates two downsides with the approach. Firstly: it's
easy to trigger a memory allocation panic but quite hard to know if
these are _actual_ bugs or false alarms. The documentation states that
`str::repeat` will panic if the capacity overflows -- sure -- but
there's always other possible reasons. Following up on research
_specifically_ for identifying integer over/underflows would be a
solid avenue here. I worry about missing part of the state space by
being shy of allocations. Second issue: writing a fuzz target for
every function in `std::str` is going to be tedious and
duplicative. Finding some way to automate this would be real spiffy.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>