Update README to reflect AFL runner (#12)
* Reduce HashMap reserve to u16

Turns out, it's still possible to crash the HashMap target even when
using the fancy bump allocator. Not sure why that is. This needs
further investigation but will, at least, allow the target to be run
again.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Update README to reflect AFL runner

This commit updates the README to reflect the changes made to use
model-based fuzzing, as opposed to straight QuickCheck. This resolves

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* 'cargo build' -> 'cargo afl build'

I mistakenly documented that one should run 'cargo build' instead of
'cargo afl build' to make the fuzz targets. This mistake was pointed
out by @ReinierMaas.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>